+++ src/main/java/org/wxl/wordTraining/model/vo/article/ArticleOneMapperVO.java
@@ -26,11 +26,11 @@ public class ArticleOneMapperVO implements Serializable {
     @ApiModelProperty("浏览量")
     private Long visitNumber;
     @ApiModelProperty("点赞数")
-    private Long praiseNumber;
+    private Integer praiseNumber;
     @ApiModelProperty("评论数")
-    private Long commentNumber;
+    private Integer commentNumber;
     @ApiModelProperty("收藏量")
-    private Long collectionNumber;
+    private Integer collectionNumber;
     @ApiModelProperty("标签")
     private String tags;
     @ApiModelProperty("封面")
