+++ src/main/java/org/wxl/wordTraining/model/vo/user/LoginUserVO.java
@@ -44,10 +44,10 @@ public class LoginUserVO implements Serializable {
     private Long pointNumber;
 
     @ApiModelProperty(value = "连续在线天数")
-    private Long coiledDay;
+    private Integer coiledDay;
 
     @ApiModelProperty(value = "在线天数")
-    private Long onlineDay;
+    private Integer onlineDay;
     @ApiModelProperty(value = "最近在线时间")
     private LocalDateTime lastLoginTime;
