+++ src/main/java/org/wxl/wordTraining/model/vo/user/UserListVO.java
@@ -43,10 +43,10 @@ public class UserListVO implements Serializable {
     private Long pointNumber;
 
     @ApiModelProperty(value = "连续在线天数")
-    private Long coiledDay;
+    private Integer coiledDay;
 
     @ApiModelProperty(value = "在线天数")
-    private Long onlineDay;
+    private Integer onlineDay;
 
     @ApiModelProperty(value = "最近在线时间")
     private LocalDateTime lastLoginTime;
